Crate solana_epoch_schedule

Configuration for epochs and slots.

Epochs mark a period of time composed of slots, for which a particular leader schedule is in effect. The epoch schedule determines the length of epochs, and the timing of the next leader-schedule selection.

The epoch schedule does not change during the life of a blockchain, though the length of an epoch does — during the initial launch of the chain there is a “warmup” period, where epochs are short, with subsequent epochs increasing in slots until they last for DEFAULT_SLOTS_PER_EPOCH.

Modules

sysvar

Structs

EpochSchedule

Constants

DEFAULT_LEADER_SCHEDULE_SLOT_OFFSET

The default number of slots before an epoch starts to calculate the leader schedule.

MAX_LEADER_SCHEDULE_EPOCH_OFFSET

The maximum number of slots before an epoch starts to calculate the leader schedule.

MINIMUM_SLOTS_PER_EPOCH

The minimum number of slots per epoch during the warmup period.